Empty dummy files to avoid warnings in tests
Up to now, most of the tests throw 'Warnings' because files cannot be found.
Indeed, those files are not needed by the respective unit test itself.
But nevertheless, in production they are required when generating TiGL code (see tigl/cpacs_gen_input/).
Hence, the code prints a warning. To avoid this, empty files with the required name are included.